requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   nltk
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
joblib==1.4.2
    # via
    #   -r requirements.in
    #   nltk
    #   scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nltk==3.8.1
    # via -r requirements.in
numpy==2.0.0
    # via
    #   pandas
    #   scikit-learn
    #   scipy
orjson==3.10.6
    # via fastapi
pandas~=1.5.3
    # via -r requirements.in
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via uvicorn
regex==2024.5.15
    # via nltk
rich==13.7.1
    # via typer
scikit-learn==1.5.1
    # via -r requirements.in
scipy==1.14.0
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
textdistance==4.6.2
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
tqdm==4.66.4
    # via nltk
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn

transformers~=4.44.0
gensim~=4.3.2
spacy~=3.7.5
pytest~=7.4.4